requirements.txt
anthropic
flask
flask-cors
gunicorn
python-dotenv
